[1014] Consolidate all qualifications code into a concern
