+++ app/models/course.rb
@@ -22,12 +22,7 @@
 #
 
 class Course < ApplicationRecord
-  enum profpost_flag: {
-    recommendation_for_qts: "",
-    professional: "PF",
-    postgraduate: "PG",
-    professional_postgraduate: "BO",
-  }
+  include WithQualifications
 
   enum program_type: {
     higher_education_programme: "HE",
@@ -78,17 +73,5 @@ class Course < ApplicationRecord
 
   def has_vacancies?
     site_statuses.with_vacancies.any?
-  end
-
-  def qualifications
-    Qualifications.new(
-      profpost_flag: profpost_flag,
-      is_pgde: pgde?,
-      is_further_education: subjects.further_education.any?,
-    ).to_a
-  end
-
-  def pgde?
-    PGDECourse.where(course_code: self.course_code, provider_code: provider.provider_code).exists?
   end
 end
