+++ app/models/pgde_course.rb
@@ -8,4 +8,10 @@
 #
 
 class PGDECourse < ApplicationRecord
+  def self.is_one?(course)
+    where(
+      course_code: course.course_code,
+      provider_code: course.provider.provider_code
+    ).exists?
+  end
 end
